requirements.txt
absl-py==1.0.0
aniso8601==9.0.1
astunparse==1.6.3
attrs==21.2.0
bleach==4.1.0
cachetools==4.2.4
certifi==2021.10.8
chardet==3.0.4
click==8.0.3
cloudinary==1.28.0
colorama==0.4.4
colorzero==2.0
decorator==5.1.0
defusedxml==0.7.1
entrypoints==0.3
Flask==2.0.2
Flask-Cors==3.0.10
Flask-RESTful==0.3.9
flatbuffers==2.0
gast==0.4.0
google-auth==2.3.3
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
gpiozero==1.6.2
grpcio==1.41.1
h5py==3.5.0
idna==2.8
ipykernel==4.6.1
ipython==5.5.0
ipython-genutils==0.2.0
itsdangerous==2.0.1
Jinja2==3.0.2
jsonschema==4.2.0
jupyter-client==7.0.6
jupyter-core==4.9.1
jupyterlab-pygments==0.1.2
keras==2.7.0
Keras-Preprocessing==1.1.2
libclang==12.0.0
Markdown==3.3.4
MarkupSafe==2.0.1
mistune==0.8.4
nbclient==0.5.4
nbconvert==6.2.0
nbformat==5.1.3
nest-asyncio==1.5.1
numpy==1.21.4
oauthlib==3.1.1
opencv-python==4.5.4.58
opt-einsum==3.3.0
ordered-set==4.0.2
packaging==21.2
pandocfilters==1.5.0
picamera==1.13
pickleshare==0.7.5
Pillow==8.4.0
portpicker==1.2.0
prompt-toolkit==1.0.18
protobuf==3.19.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.10.0
PyLaTeX==1.4.1
pyparsing==2.4.7
pyrsistent==0.18.0
pytesseract==0.3.8
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
pywin32==302
pyzmq==22.3.0
requests==2.21.0
requests-oauthlib==1.3.0
rsa==4.7.2
simplegeneric==0.8.1
six==1.12.0
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow-cpu==2.7.0
tensorflow-estimator==2.7.0
tensorflow-io-gcs-filesystem==0.22.0
termcolor==1.1.0
testpath==0.5.0
tornado==4.5.3
traitlets==5.1.1
typing-extensions==3.10.0.2
urllib3==1.24.3
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.2
wrapt==1.13.3
